Introducing a new BaseObject as a parent class for entities and (data) structures.
