+++ src/Entity/Entity.php
@@ -19,6 +19,7 @@
 namespace Apigee\Edge\Entity;
 
 use Apigee\Edge\Api\Management\Entity\Organization;
+use Apigee\Edge\Structure\BaseObject;
 
 /**
  * Base representation of an Edge entity.
@@ -34,57 +35,12 @@ use Apigee\Edge\Api\Management\Entity\Organization;
  *   additional helper methods that make developers life easier. @see Organization::isCpsEnabled()
  * - Entity properties with object or array types must be initialized.
  */
-class Entity implements EntityInterface
+abstract class Entity extends BaseObject implements EntityInterface
 {
     /**
      * On the majority of entities this property is the primary entity.
      */
     private const DEFAULT_ID_FIELD = 'name';
-
-    /**
-     * Entity constructor.
-     *
-     * @param array $values
-     *   Associative array with entity properties and their values.
-     *
-     * @throws \ReflectionException
-     */
-    public function __construct(array $values = [])
-    {
-        $ro = new \ReflectionObject($this);
-        foreach ($ro->getProperties() as $property) {
-            if (!array_key_exists($property->getName(), $values)) {
-                continue;
-            }
-            $setter = 'set' . ucfirst($property->getName());
-            if ($ro->hasMethod($setter)) {
-                $value = $values[$property->getName()];
-                $rm = new \ReflectionMethod($this, $setter);
-                $rm->invoke($this, $value);
-            }
-        }
-    }
-
-    /**
-     * Deep clone for entity structures.
-     *
-     * Inspired by https://github.com/kore/DataObject/blob/master/src/Kore/DataObject/DataObject.php
-     */
-    public function __clone()
-    {
-        $ro = new \ReflectionObject($this);
-        foreach ($ro->getProperties() as $property) {
-            $property->setAccessible(true);
-            $value = $property->getValue($this);
-            if (is_object($value)) {
-                $this->{$property->getName()} = clone $value;
-            }
-            if (is_array($value)) {
-                $this->cloneArray($value);
-                $this->{$property->getName()} = $value;
-            }
-        }
-    }
 
     /**
      * @inheritdoc
@@ -100,22 +56,5 @@ class Entity implements EntityInterface
     public function idProperty(): string
     {
         return self::DEFAULT_ID_FIELD;
-    }
-
-    /**
-     * Deep clone for arrays.
-     *
-     * @param array $array
-     */
-    private function cloneArray(array &$array): void
-    {
-        foreach ($array as $key => $value) {
-            if (is_object($value)) {
-                $array[$key] = clone $value;
-            }
-            if (is_array($value)) {
-                $this->cloneArray($array[$key]);
-            }
-        }
     }
 }
